feat(data-service): import KPT unpack GeometryCollection
Related work items: #2264, #2343

+++ data-service/src/main/java/ru/mycrg/data_service/service/cqrs/table_records/handlers/DeleteMultipleTableRecordsHandler.java
@@ -37,9 +37,12 @@ public class DeleteMultipleTableRecordsHandler implements IRequestHandler<Delete
         ResourceQualifier rQualifier = request.getrQualifiers();
 
         try {
-            ddlTablesBase.isExist(rQualifier);
-            spatialRecordsDao.removeMultipleRecords(rQualifier, request.getIds());
-            removeRecordsFromExtensionTableByObjectId(rQualifier, request.getIds());
+            if (ddlTablesBase.isExist(rQualifier)) {
+                spatialRecordsDao.removeMultipleRecords(rQualifier, request.getIds());
+                removeRecordsFromExtensionTableByObjectId(rQualifier, request.getIds());
+            } else {
+                throw new DataServiceException("Не найдена таблица: " + rQualifier.getQualifier());
+            }
         } catch (CrgDaoException e) {
             throw new DataServiceException(e.getMessage(), e.getCause());
         }
